observability/langsmith/google-adk/code-debug-agent-demo/pyproject.toml
[project]
name = "code-debug-agent-demo-langsmith"
version = "0.1.0"
description = "Google ADK agent for debugging assistance using Stack Exchange (with LangSmith tracing)"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Code Debug Agent Team", email = "team@example.com"},
]
dependencies = [
    "google-adk>=1.0.0",
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",
    "google-genai>=1.9.0",
    "pydantic>=2.10.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "beautifulsoup4>=4.12.0",
    "stackapi>=0.3.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.0.0",
    "black>=24.10.0",
    "ruff>=0.9.0",
    "tenacity>=9.0.0",
    "langsmith>=0.1.147",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["E501", "B008"]
